fix(windows): Fix "Access is defined." error on Windows when renaming.
